+++ src/AppManager.ts
