+++ src/abstraction/IProtocol.ts
