+++ src/abstraction/electron/protocol.ts
