+++ src/abstraction/protocol.ts
