+++ src/api/IPackageEntry.ts
