+++ src/electron/Dialog.ts
