+++ src/electron/ui/splash.html
