+++ src/lib/WritableMemoryStream.ts
