+++ src/lib/hashes.ts
